# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eoogle Analytics",
        "version": "1.0.0"
    },
    "paths": {
        "/api/analytics/hits/by-link/{link}": {
            "get": {
                "tags": [
                    "hits",
                    "analytics"
                ],
                "summary": "getHitsByLink",
                "operationId": "getHitsByLink",
                "parameters": [
                    {
                        "name": "link",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Statistics"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/hits//by-type/{linkType}": {
            "get": {
                "tags": [
                    "hits"
                ],
                "summary": "getHitsByLinkInInterval",
                "operationId": "getHitsByLinkInInterval",
                "parameters": [
                    {
                        "name": "linkType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/user-journey/{customer_uuid}": {
            "get": {
                "tags": [
                    "user-journey"
                ],
                "summary": "getUserJourney",
                "operationId": "getUserJourney",
                "parameters": [
                    {
                        "name": "customer_uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/user-journey/{customer_uuid}/similar": {
            "get": {
                "tags": [
                    "user-journey"
                ],
                "summary": "getUsersJourneyByUserIdentifier",
                "operationId": "getUsersJourneyByUserIdentifier",
                "parameters": [
                    {
                        "name": "customer_uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/statistics": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "link": {
                                        "type": "string",
                                        "example": "google.com"
                                    },
                                    "link_type": {
                                        "type": "string",
                                        "example": "homepage"
                                    },
                                    "customer_uuid": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "09bfb8dc-fdc9-3bc0-a417-2eaf1f293c13"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Statistics": {
                "title": "Statistics",
                "properties": {
                    "link": {
                        "title": "link",
                        "type": "string"
                    },
                    "link_type": {
                        "title": "link_type",
                        "type": "string"
                    },
                    "customer_uuid": {
                        "title": "customer_uuid",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}